# api_spec.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "RASO Platform API",
                 "description":  "Research paper Automated Simulation \u0026 Orchestration Platform",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/jobs":  {
                                    "get":  {
                                                "summary":  "List Jobs",
                                                "description":  "List all jobs.",
                                                "operationId":  "list_jobs_api_jobs_get",
                                                "responses":  {
                                                                  "200":  {
                                                                              "description":  "Successful Response",
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {

                                                                                                                                  }
                                                                                                                   }
                                                                                          }
                                                                          }
                                                              }
                                            },
                                    "post":  {
                                                 "summary":  "Submit Job",
                                                 "description":  "Submit a new video generation job.",
                                                 "operationId":  "submit_job_api_jobs_post",
                                                 "requestBody":  {
                                                                     "content":  {
                                                                                     "application/json":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/JobRequest"
                                                                                                                         }
                                                                                                          }
                                                                                 },
                                                                     "required":  true
                                                                 },
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "Successful Response",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/JobResponse"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           },
                                                                   "422":  {
                                                                               "description":  "Validation Error",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           }
                                                               }
                                             }
                                },
                  "/api/jobs/{job_id}":  {
                                             "get":  {
                                                         "summary":  "Get Job Status",
                                                         "description":  "Get job status.",
                                                         "operationId":  "get_job_status_api_jobs__job_id__get",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "job_id",
                                                                                "in":  "path",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Job Id"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "Successful Response",
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/JobStatus"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                           "422":  {
                                                                                       "description":  "Validation Error",
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   }
                                                                       }
                                                     }
                                         },
                  "/api/jobs/{job_id}/download":  {
                                                      "get":  {
                                                                  "summary":  "Download Video",
                                                                  "description":  "Download generated video.",
                                                                  "operationId":  "download_video_api_jobs__job_id__download_get",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "job_id",
                                                                                         "in":  "path",
                                                                                         "required":  true,
                                                                                         "schema":  {
                                                                                                        "type":  "string",
                                                                                                        "title":  "Job Id"
                                                                                                    }
                                                                                     }
                                                                                 ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "Successful Response",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {

                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            },
                                                                                    "422":  {
                                                                                                "description":  "Validation Error",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {
                                                                                                                                                        "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            }
                                                                                }
                                                              }
                                                  }
              },
    "components":  {
                       "schemas":  {
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "JobRequest":  {
                                                          "properties":  {
                                                                             "paper_input":  {
                                                                                                 "$ref":  "#/components/schemas/PaperInput"
                                                                                             },
                                                                             "options":  {
                                                                                             "anyOf":  [
                                                                                                           {
                                                                                                               "additionalProperties":  true,
                                                                                                               "type":  "object"
                                                                                                           },
                                                                                                           {
                                                                                                               "type":  "null"
                                                                                                           }
                                                                                                       ],
                                                                                             "title":  "Options"
                                                                                         }
                                                                         },
                                                          "type":  "object",
                                                          "required":  [
                                                                           "paper_input"
                                                                       ],
                                                          "title":  "JobRequest",
                                                          "description":  "Job submission request."
                                                      },
                                       "JobResponse":  {
                                                           "properties":  {
                                                                              "job_id":  {
                                                                                             "type":  "string",
                                                                                             "title":  "Job Id"
                                                                                         },
                                                                              "status":  {
                                                                                             "type":  "string",
                                                                                             "title":  "Status"
                                                                                         },
                                                                              "created_at":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "date-time",
                                                                                                 "title":  "Created At"
                                                                                             }
                                                                          },
                                                           "type":  "object",
                                                           "required":  [
                                                                            "job_id",
                                                                            "status",
                                                                            "created_at"
                                                                        ],
                                                           "title":  "JobResponse",
                                                           "description":  "Job submission response."
                                                       },
                                       "JobStatus":  {
                                                         "properties":  {
                                                                            "job_id":  {
                                                                                           "type":  "string",
                                                                                           "title":  "Job Id"
                                                                                       },
                                                                            "status":  {
                                                                                           "type":  "string",
                                                                                           "title":  "Status"
                                                                                       },
                                                                            "progress":  {
                                                                                             "type":  "number",
                                                                                             "title":  "Progress"
                                                                                         },
                                                                            "current_agent":  {
                                                                                                  "anyOf":  [
                                                                                                                {
                                                                                                                    "type":  "string"
                                                                                                                },
                                                                                                                {
                                                                                                                    "type":  "null"
                                                                                                                }
                                                                                                            ],
                                                                                                  "title":  "Current Agent"
                                                                                              },
                                                                            "created_at":  {
                                                                                               "type":  "string",
                                                                                               "format":  "date-time",
                                                                                               "title":  "Created At"
                                                                                           },
                                                                            "updated_at":  {
                                                                                               "type":  "string",
                                                                                               "format":  "date-time",
                                                                                               "title":  "Updated At"
                                                                                           },
                                                                            "error_message":  {
                                                                                                  "anyOf":  [
                                                                                                                {
                                                                                                                    "type":  "string"
                                                                                                                },
                                                                                                                {
                                                                                                                    "type":  "null"
                                                                                                                }
                                                                                                            ],
                                                                                                  "title":  "Error Message"
                                                                                              },
                                                                            "result":  {
                                                                                           "anyOf":  [
                                                                                                         {
                                                                                                             "additionalProperties":  true,
                                                                                                             "type":  "object"
                                                                                                         },
                                                                                                         {
                                                                                                             "type":  "null"
                                                                                                         }
                                                                                                     ],
                                                                                           "title":  "Result"
                                                                                       }
                                                                        },
                                                         "type":  "object",
                                                         "required":  [
                                                                          "job_id",
                                                                          "status",
                                                                          "progress",
                                                                          "created_at",
                                                                          "updated_at"
                                                                      ],
                                                         "title":  "JobStatus",
                                                         "description":  "Job status response."
                                                     },
...
                                   }
                   }
}